Zoom to selected is stored as a URL query
When clicking "Zoom to selected" we now add a URL query
`treeZoom=selected`, which is cleared upon the next zoom action.
This allows us to load a page to a zoomed in view of the tree.
This behavior is similar to when zooming to a branch which a clade
label, which results in a query such as `label=clade:A1`. In the
situation where both are applicable, the clade label takes
priority.

Note that due to the current implementation of zooming in Auspice,
"zoom to selected" will zoom to different subtrees. As an example,
load each of the following datasets and click "zoom to selected":
`/flu/seasonal/h3n2/ha/3y?f_clade_membership=A1b/137F,A1b/197R`,
`/flu/seasonal/h3n2/ha/3y?f_clade_membership=A1b/137F,A1b/197R&label=clade:A1b/131K`.
The resulting subtrees will be different for each case, however both
result in a URL query `treeZoom=selected`. This is a shortcoming
of the implementation used in this PR, however one that is not easily
solved.